+++ cardano-tracer/CHANGELOG.md
@@ -1,5 +1,9 @@
 # ChangeLog
 
+## 0.2.1
+
+* Updated to `ouroboros-network-0.10`
+
 ## 0.2.0
 
 * Updated to `ouroboros-network-0.9.1`